Каунтер в фильтрах через запрос и join для локации

+++ app/src/main/kotlin/com/itrocket/union/filter/presentation/view/FilterScreen.kt
@@ -119,7 +119,7 @@ private fun FilterContent(
 
 @Composable
 private fun FilterBottomBar(
-    resultCount: Int,
+    resultCount: Long,
     onBtnClickListener: () -> Unit,
     resources: Resources
 ) {
